Detect failed boot attempts
Set CONFIG_BOOTCOUNT_LIMIT=1.

Set CONFIG_SPL_FS_FAT=n to make space space in the SPL.

After an update is applied, set the U-Boot environment variables
upgrade_available to 1, and bootcount to 0.

When U-Boot starts, if upgrade_available is different from 0, bootcount
is incremented and saved.

If bootlimit is defined and bootcount > bootlimit, U-Boot will run the
commands in altbootcmd instead of bootcmd.

When Linux has booted successfully, a userspace application can take care
of setting upgrade_available and bootcount back to 0.

Otherwise, the system will be rebooted either manually (pressing reset)
or automatically (if a watchdog is used), and bootcount will increase
until bootlimit is reached.